requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.10
install==1.3.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
numpy==1.19.0
pandas==1.0.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==1.0.1
